requirements.txt

python-telegram-bot==13.15
pandas
numpy
ta
scikit-learn
joblib
requests
upstox-python-sdk
bybit
websocket-client
schedule
